remove unnecessary int cast

+++ src/test/java/com/natelenergy/porter/worker/AvroReaderTest.java
@@ -39,7 +39,7 @@ public class AvroReaderTest {
     System.out.println( "AFTER: "+ JSONHelper.toJSON(last.getDB(null)));
     
 
-    assertThat( last.get("fINT").getValue() ).isEqualTo( (int)10 );
+    assertThat( last.get("fINT").getValue() ).isEqualTo( 10 );
     assertThat( last.get("fLONG").getValue() ).isEqualTo( (long)10 );
     assertThat( last.get("fFLOAT").getValue() ).isEqualTo( (float)10 );
     assertThat( last.get("fDOUBLE").getValue() ).isEqualTo( (double)10 );
